ethereum/burn_test.go
package ethereum

import (
	"context"
	"math/big"
	"testing"
	"time"

	"github.com/dan13ram/wpokt-backend/app"
	"github.com/ethereum/go-ethereum/accounts/abi/bind"
	"github.com/ethereum/go-ethereum/common"
	"github.com/ethereum/go-ethereum/core/types"
	"github.com/ethereum/go-ethereum/ethclient"
)

// Mock implementation of the Client interface
type mockClient struct{}

func (c *mockClient) GetBlockNumber() (uint64, error) {
	return 10, nil
}

func (c *mockClient) GetClient() *ethclient.Client {
	return nil
}

func (c *mockClient) GetChainId() (*big.Int, error) {
	return nil, nil
}

func (c *mockClient) ValidateNetwork() {}

type mockBurnAndBridgeIterator struct{}

func (i *mockBurnAndBridgeIterator) Next() bool {
	return false
}

func (i *mockBurnAndBridgeIterator) Event() *WrappedPocketBurnAndBridge {
	return &WrappedPocketBurnAndBridge{
		Raw: types.Log{
			BlockNumber: 1,
			TxHash:      common.HexToHash("test-hash"),
			Index:       1,
		},
		From:        common.HexToAddress("sender-address"),
		PoktAddress: common.HexToAddress("recipient-address"),
		Amount:      big.NewInt(100),
	}
}

// Mock implementation of the WrappedPocket contract
type mockWrappedPocket struct{}

func (c *mockWrappedPocket) FilterBurnAndBridge(opts *bind.FilterOpts, _amount []*big.Int, _from []common.Address, _poktAddress []common.Address) (BurnAndBridgeIterator, error) {
	return &mockBurnAndBridgeIterator{}, nil
}

func setupBurnMonitor() *WPOKTBurnMonitor {
	app.DB = &mockDB{}
	Client = &mockClient{}

	monitor := &WPOKTBurnMonitor{
		stop:               make(chan bool),
		startBlockNumber:   1,
		currentBlockNumber: 10,
		monitorInterval:    10 * time.Second,
		wpoktContract:      &mockWrappedPocket{},
	}

	return monitor
}

func TestWPOKTBurnMonitor_Stop(t *testing.T) {
	monitor := setupBurnMonitor()

	// Call the Stop() method
	go func() {
		time.Sleep(2 * time.Second) // Wait for a few seconds before stopping
		monitor.Stop()
	}()

	// Start the monitor and wait for it to stop
	go monitor.Start()

	// If the monitor doesn't stop within a reasonable time, it means Stop() didn't work
	select {
	case <-monitor.stop:
		// Monitor stopped as expected
	case <-time.After(5 * time.Second):
		t.Errorf("Stop() did not stop the monitor")
	}
}

func TestWPOKTBurnMonitor_UpdateCurrentBlockNumber(t *testing.T) {
	monitor := &WPOKTBurnMonitor{}

	// Mock the Client to return a fixed block number
	Client = &mockClient{}

	// Call the UpdateCurrentBlockNumber() method
	monitor.UpdateCurrentBlockNumber()

	// Check if the currentBlockNumber field is updated correctly
	if monitor.currentBlockNumber != 10 {
		t.Errorf("UpdateCurrentBlockNumber() failed. Expected currentBlockNumber: 10, got: %d", monitor.currentBlockNumber)
	}
}

func TestWPOKTBurnMonitor_HandleBurnEvent(t *testing.T) {
	monitor := &WPOKTBurnMonitor{}

	event := &WrappedPocketBurnAndBridge{
		Raw: types.Log{
			BlockNumber: 1,
			TxHash:      common.HexToHash("test-hash"),
			Index:       1,
		},
		From:        common.HexToAddress("sender-address"),
		PoktAddress: common.HexToAddress("recipient-address"),
		Amount:      big.NewInt(100),
	}

	// Mock the InsertOne function to return a nil error
	app.DB = &mockDB{}

	// Call the HandleBurnEvent() method
	result := monitor.HandleBurnEvent(event)

	// Check if the method returns the expected result
	if result != true {
		t.Errorf("HandleBurnEvent() failed. Expected: true, got: %t", result)
	}
}

func TestWPOKTBurnMonitor_SyncBlocks(t *testing.T) {
	monitor := &WPOKTBurnMonitor{}

	// Mock the WrappedPocket contract to return a nil error
	monitor.wpoktContract = &mockWrappedPocket{}

	// Call the SyncBlocks() method
	result := monitor.SyncBlocks(1, 10)

	// Check if the method returns the expected result
	if result != true {
		t.Errorf("SyncBlocks() failed. Expected: true, got: %t", result)
	}
}

func TestWPOKTBurnMonitor_SyncTxs(t *testing.T) {
	monitor := setupBurnMonitor()

	// Call the SyncTxs() method
	result := monitor.SyncTxs()

	// Check if the method returns the expected result
	if result != true {
		t.Errorf("SyncTxs() failed. Expected: true, got: %t", result)
	}
}

func TestWPOKTBurnMonitor_Start(t *testing.T) {
}

func TestNewBurnMonitor(t *testing.T) {
	currentBlockNumber := uint64(10)

	// Mock the Config values
	app.Config.Ethereum.WPOKTContractAddress = "0xabc123"
	app.Config.Ethereum.StartBlockNumber = -1
	app.Config.Ethereum.MonitorIntervalSecs = 10

	// Call the NewBurnMonitor() method
	monitor := NewBurnMonitor().(*WPOKTBurnMonitor)

	// Check if the monitor is initialized correctly
	if monitor.wpoktContract == nil {
		t.Errorf("NewBurnMonitor() failed. wpoktContract is nil")
	}

	if monitor.startBlockNumber != currentBlockNumber {
		t.Errorf("NewBurnMonitor() failed. startBlockNumber is not updated correctly")
	}

	if monitor.currentBlockNumber != currentBlockNumber {
		t.Errorf("NewBurnMonitor() failed. currentBlockNumber is not updated correctly")
	}

	expectedInterval := 10 * time.Second
	if monitor.monitorInterval != expectedInterval {
		t.Errorf("NewBurnMonitor() failed. Expected monitorInterval: %s, got: %s", expectedInterval, monitor.monitorInterval)
	}
}

// Mock implementation of the database
type mockDB struct{}

func (db *mockDB) InsertOne(collectionName string, document interface{}) error {
	return nil
}

func (db *mockDB) Connect(ctx context.Context) error {
	return nil
}

func (db *mockDB) SetupIndexes() error {
	return nil
}

func (db *mockDB) Disconnect() error {
	return nil
}
